backend/requirements.txt
fastapi==0.115.7
uvicorn[standard]==0.34.0
python-multipart==0.0.20
pydantic==2.10.6
python-dotenv==1.0.1
pypdf==5.1.0
python-docx==1.1.2
openai==1.61.0
httpx==0.28.1
